Rename websocket to wsActionClient
