+++ src/Kernel/actionClient.js
@@ -1,16 +1,19 @@
+/** Exports an event emitter that provides ipcRenderer functionality on web and electron-renderer targets */
+
+let actionClient
 if (process.env.WEBPACK_TARGET === 'web') {
-  const { createActionClient } = require('../websocket')
-  const backendHost = `ws://${location.host}/wss`
-  const ws = new WebSocket(backendHost)
+  const { createActionClient } = require('../wsActionClient')
+  const backendHost = `ws://${window.location.host}/wss`
+  const ws = new window.WebSocket(backendHost)
   ws.addEventListener('error', (e) => console.error(`WebSocket error connecting to backend ${backendHost}`, e))
   ws.addEventListener('open', () => console.log(`Opened WebSocket connection to backend ${backendHost}`))
   ws.addEventListener('close', () => console.log(`Closed WebSocket connection to backend ${backendHost}`))
 
-  module.exports = createActionClient(ws)
+  actionClient = createActionClient(ws)
 } else if (process.env.WEBPACK_TARGET === 'electron-renderer') { // electron-renderer
-  const actionClient = require('electron').ipcRenderer
-  actionClient.emit('open')
-  module.exports = actionClient
+  actionClient = require('electron').ipcRenderer
 } else {
-  module.exports = {}
+  actionClient = {}
 }
+
+module.exports = actionClient
